Create a components directory where we create a new file that contains the navbar that its imported into app.js.
